update ts demo , lỗi model
