+++ .gitignore
@@ -1,4 +1,3 @@
 node_modules
 .env
-*.json
 package-lock.json
