+++ package.json
@@ -3,13 +3,12 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
-  "type": "module",
   "engines": {
     "node": ">=16.9"
   },
   "scripts": {
     "build": "npm install",
-    "start": "nodemon src/index.js",
+    "start": "nodemon src/index.ts",
     "lint": "npx eslint ./src && npx prettier ./src",
     "lint-fix": "npx eslint --fix ./src && npx prettier --write ./src"
   },
@@ -29,6 +28,13 @@
   },
   "devDependencies": {
     "eslint": "^8.6.0",
-    "eslint-config-prettier": "^8.3.0"
+    "eslint-config-prettier": "^8.3.0",
+    "@types/cors": "^2.8.13",
+    "@types/express": "^4.17.17",
+    "@types/node": "^18.13.0",
+    "concurrently": "^7.6.0",
+    "nodemon": "^2.0.20",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5"
   }
 }
