+++ src/commands/milk.ts
@@ -1,20 +1,24 @@
-import config from '../config/config.js';
-import roundDouble from '../helpers/roundDouble.js';
-import randomRange from '../helpers/randomRange.js';
+import config from '../config/config';
+import roundDouble from '../helpers/roundDouble';
+import randomRange from '../helpers/randomRange';
 import {
 	decStrength,
 	getUser,
 	getTotalMilkByDay,
 	updateUser,
 	addUser,
-} from '../controllers/userController.js';
+} from '../controllers/userController';
+import { Message } from 'discord.js';
+import { User } from '../models/userModel';
+import mongoose from 'mongoose';
+import {MilkTankModel} from '../models/milkTankModel';
 
-const milk = async (message) => {
+const milk = async (message: Message) => {
 	const user = await getUser(message.author.id),
 		milk = roundDouble(randomRange(config.maxMilk, config.minMilk));
 	if (user) {
 		const newStrength = await decStrength(user),
-			diffTime = user.lastTimeTakeMilk - new Date(),
+			diffTime = new Date(user.lastTimeTakeMilk).getTime() - new Date().getTime(),
 			diffSecond = Math.abs(Math.ceil(diffTime / 1000)),
 			timeLeft = config.coolDownMilk - diffSecond;
 
@@ -58,7 +62,10 @@ const milk = async (message) => {
 			}
 		}
 	} else {
-		const newUser = {
+		const newUser: User = {
+			_id: new mongoose.Types.ObjectId(),
+			userTagName: message.author.tag,
+			numberOfCow: 1,
 			cow: {
 				strength: 100,
 				dateOfBirth: new Date(),
@@ -66,7 +73,7 @@ const milk = async (message) => {
 			},
 			userId: message.author.id,
 			lastTimeTakeMilk: new Date(),
-			milkTank: [{ milk, takingTime: new Date() }],
+			milkTank: [new MilkTankModel({ milk, takingTime: new Date() })],
 			totalMilk: milk,
 		};
 		await addUser(newUser);
