--- src/helpers/randomRange.js
@@ -1,1 +1,0 @@
-export default (min, max) => Math.random() * (max - min) + min;
