+++ src/helpers/randomRange.ts
@@ -1,0 +1,1 @@
+export default (min: number, max: number) => Math.random() * (max - min) + min;
