+++ src/helpers/removeTag.ts
@@ -1,4 +1,4 @@
-export default function (name) {
+export default function (name: string) {
     let nameParts = name.split('#');
     if (nameParts.length == 1) return name;
     if (parseInt(nameParts[1]) == 0)
