--- src/helpers/roundDouble.js
@@ -1,1 +1,0 @@
-export default (number) => Math.round(number * 100) / 100;
