+++ src/helpers/roundDouble.ts
@@ -1,0 +1,1 @@
+export default (number: number) => Math.round(number * 100) / 100;
