--- src/models/milkTankModel.js
@@ -1,8 +1,0 @@
-import mongoose from 'mongoose';
-const { Schema } = mongoose,
-	MilkTank = new Schema({
-		milk: { type: Number, default: 0 },
-		takingTime: Date,
-	});
-
-export default MilkTank;
