+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "compilerOptions": {
+      "module": "commonjs",
+      "target": "es6",
+      "rootDir": "./",
+      "esModuleInterop": true
+    }
+  }
